Stub initial RxWriter support and reframe Store concept to be immutable
I think the right design is to treat each RxStore as immutable, and the
result of a RxWriter changeset produces a new instance.  This makes its
much easier for each thread to maintain a consistent copy (even if out
of date).  And it becomes a local matter how you update the "master"
reference of the store (if even applicable)